requirements.txt
colorzero==2.0
gpiozero==2.0.1
lgpio==0.2.2.0
paho-mqtt==2.1.0

